Manage dnf-query.py as a Cookbook File.
There is a bug in the current version of Test Kitchen which causes the
`dnf-query.py` script to be delete during the cleanup. (See
test-kitchen/test-kitchen#929 for details.) This changes the way
the file is dropped by using the Chef `cookbook_file` resource, and
places the file in a common path.